Use absolute source map and index.js bundle paths
With the absolute source map paths the source maps show up in Chrome Dev
Tools as different files than the served bundles, which allows the
bundle names to be 'index.js' again, making hot reloading work for these
files. For hot reloading of submodules some other solution will have to
be devised, but for the moment this improves the dev workflow.