Do not polyfill library transformations
In order to publish the individual libraries more easily, prepare the
webpack configuration for use by the libraries, in addition to the demo.
Rather than adding polyfills through the babel runtime transformation
plugin, which results in imperfect polyfilling of some instance methods,
expect that applications will polyfill their environments. This decision
seems reasonable especially in light of the switch to using only shipped
proposals in the environment preset that avoids requiring any consumers
to apply highly experimental transformations.

Instead, add an import of @babel/polyfill to the demo and let the env
preset transform it to a minimal polyfill for the target environments.