Move babel runtime dependency into packages
The workspace is really a development thing, so it doesn't need any
dependencies outside of dev dependencies. The packages themselves will
need the runtime at runtime because the transform will result in helpers
being pulled from it. However, it's nice to let consumers use the
version they want by specifying it as a peer dependency.