Remove uses of require from validation tests
Strict .mjs does not have access to `require`. Due to a bug in `esm`, it
worked up until now. However, removing it is necessary to unblock
upgrading `esm`.

At the same time, address the fact that it is not ideal to assume that
the node modules directory is relative to the current working directory.
Instead, rely on `resolve` to do a recursive search.