Simplify refinement implementation
Implement refinement naively by running the refining selector against
the original scope and constraining the valid matches to those that are
within the initial match. This approach removes the need to compose the
initial match and the refining match to produce a match that can be
yielded to the caller. More importantly, it removes the need to use the
initial match as a scope for the refinement, which makes the
implementation ignorant of valid scope types.