Remove esm and run tests in a cjs environment
Remove the complexity of using esm to support ES module syntax in the
test environment.

The esm module no longer seems to be keeping up with bleeding edge
changes to Node v13 module semantics and it's common in the ecosystem
still to transform code to CommonJS modules for testing as it makes it
possible to mock exports.