Put build output in /web/dist
It’s a rather hacky configuration, as webpack appears unwilling to just
put all static files in the right place. It needs explicit entrypoints
now, and insists on creating a bogus index/index.js. So be it.

We could try Parcel some time instead, which supposedly handles html
files as entry points out of the box.