More tests — some failing
Three tests fail because of an opinionated expectation: I expect a match
at the start of a the text node to have that textNode as start container
rather than its parent element (although technically such a Range should
be equivalent).

Two tests fail because we really mess up when the ancestorContainer of
the scope contains text before its startContainer. Ouch.