Some fixes to satisfy tests
Always give a prefix and suffix, as recommended in the spec:
  “Each TextQuoteSelector SHOULD have exactly 1 prefix property”
…so now it is an empty string when no prefix/suffix is needed.
I am not sure if this is valuable or a waste a bytes. I’d be equally
happy to revert this behaviour.

Fix mistakes leading to needless prefix or suffix in some cases, and
a lack of them when selecting the first or last characters (partly
regressions from c94ccda and 8b29fec, or already faulty before that).