Raise the TypeScript target to es2018
Node.js v10 has full support for ECMAScript 2018 features, with only the
exception of not throwing TypeError exceptions when a Proxy object
returns an Array having duplicated entries from its ownKeys method. This
target is consistent with that used by the recommended node10 base
config.

Also set the parser options for the build and configuration scripts.