Make rangeSelector work as a plugin
This is why the plugin/middleware is needed: rangeSelector needs to
access the outermost matcherCreator function, so that its inner
selectors can also be refined.

Range implementation is currently DOM-specific, so I suppose it has to
be a plugin from the dom package. Not sure if we will be able to change
that any time soon. (would require abstracting Range, I suppose..)
A downside is that types are a mess; luckily the demo is not yet in TS,
but I expect many red squiggly lines..