Change approach to choose prefix/suffix.
This makes it faster, but does not guarantee to find the minimum number
of characters needed to disambiguate. I think it should be at most twice
that minimum though.

Note that the minimum is in any case not what one wants if the goal is
to make the selector robust to small document changes. We should add a
less stingy approach some day.