Remove the import/no-internal-modules lint rule
With Node.js adopting a standard for packages to declare their exports,
this rule may not be useful anymore. It should be legal to import any
valid export from a package.

An open [issue] on eslint-plugin-import tracks this ecosystem feature
and when that resolves there may be a better rule to enable.

[issue]: https://github.com/benmosher/eslint-plugin-import/issues/1868