Create release clones without tags
If the release clone is allowed to contain tags, it will contain the
release candidate tag, which may be confusing to someone looking at a
release tarball. Instead, clone without tags, and instructions for
verifying a release can include verifying that the remote tag matches
the state of the release.