Remove main export for the apache-annotator meta-package
Reduce the number of different ways to import modules by removing the
main export for the apache-annotator meta-package.

Consumers can import from apache-annotator/<package> or use the scoped
packages in @apache-annotator/<package>, but not from named exports of a
main module for the meta-package.

This setup provides a nice symmetry to the import styles.