Opt into explicit TypeScript index property access
Opt into a new feature of TypeScript that restores an old behavior
wherein accessing properties that match an index signature requires
using bracket syntax. With this change, TypeScript will catch dotted
attributed access that might be incorrect, but allow access with bracket
notation that may be useful in certain scenarios.