Move SPDX header comments and remove copyright dates
Keep the SPDX header comments, but move them underneath the ASF headers
to make files human-friendly first and machine-friendly second.

Remove copyright dates from header comments. The ASF recommends against
adding an ASF copyright in the header comment, but does not forbid it.
Keep the copyright headers for compliance with the REUSE specification,
but remove the dates to reduce maintenance burder.